before swagger

+++ src/main/java/org/ohgiraffers/luluproject/domain/Post.java
@@ -13,7 +13,7 @@ public class Post{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int post_id;
+    private Long post_id;
 
     @Column(length = 50, nullable = false)
     private String title;
